requirements.txt
asgiref==3.4.1    
certifi==2021.10.8
charset-normalizer==2.0.9    
Django==3.2.9    
django-cors-headers==3.10.0   
django-filter==21.1
django-phonenumber-field==6.0.0
django-truncate==0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
environ==1.0
idna==3.3
phonenumbers==8.12.40
Pillow==8.4.0
pip==21.3.1
psycopg2==2.9.2
PyJWT==2.3.0
pytz==2021.3
razorpay==1.2.0
requests==2.26.0
setuptools==39.0.1
sqlparse==0.4.2
stripe==2.64.0
typing_extensions==4.0.0
urllib3==1.26.7
